refactor(Client): 패키징 수정
