+++ src/main/java/step3and4/client/input/InputView.java
@@ -1,4 +1,4 @@
-package step3and4.input;
+package step3and4.client.input;
 
 import java.io.IOException;
