+++ src/main/java/step3and4/client/output/OutputView.java
@@ -1,4 +1,4 @@
-package step3and4.output;
+package step3and4.client.output;
 
 @FunctionalInterface
 public interface OutputView {
