+++ src/main/java/step3and4/game/car/Cars.java
@@ -1,7 +1,7 @@
 package step3and4.game.car;
 
-import step3and4.number.Number;
-import step3and4.output.OutputView;
+import step3and4.client.number.Number;
+import step3and4.client.output.OutputView;
 
 import java.util.List;
 import java.util.Objects;
