+++ src/main/java/step3and4/game/car/Position.java
@@ -1,6 +1,6 @@
 package step3and4.game.car;
 
-import step3and4.output.OutputView;
+import step3and4.client.output.OutputView;
 
 import java.util.Objects;
