+++ src/test/java/step3and4/game/car/CarsTest.java
@@ -2,9 +2,9 @@ package step3and4.game.car;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import step3and4.number.Number;
-import step3and4.number.RandomNumber;
-import step3and4.output.OutputView;
+import step3and4.client.number.Number;
+import step3and4.client.number.RandomNumber;
+import step3and4.client.output.OutputView;
 
 import java.util.List;
